# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/chat": {
      "post": {
        "operationId": "ChatController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Chat"
        ]
      },
      "get": {
        "operationId": "ChatController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/{id}": {
      "get": {
        "operationId": "ChatController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Chat"
        ]
      },
      "patch": {
        "operationId": "ChatController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChatDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Chat"
        ]
      },
      "delete": {
        "operationId": "ChatController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/ask-assistant/{id}": {
      "post": {
        "description": "Ask the assistant a question and get a response",
        "operationId": "ChatController_askAssistant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponseDto"
                }
              }
            }
          }
        },
        "summary": "Ask the assistant a question",
        "tags": [
          "Chat"
        ]
      }
    },
...
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sets access_token and refresh_token cookies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Sign in with email and password",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "summary": "Get current user from access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Issues new tokens (also set as HttpOnly cookies)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponseDto"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "summary": "Refresh access token using refresh_token cookie",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/{id}": {
      "get": {
        "operationId": "AuthController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      },
      "patch": {
        "operationId": "AuthController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      },
      "delete": {
        "operationId": "AuthController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Converse AI API",
    "description": "Enterprise knowledge assistant API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateChatDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateChatDto": {
        "type": "object",
        "properties": {}
      },
      "ChatSourceDto": {
        "type": "object",
        "properties": {
          "citationNumber": {
            "type": "number"
          },
          "matchScore": {
            "type": "number"
          },
          "fileName": {
            "type": "string"
          },
          "cloudinaryUrl": {
            "type": "string"
          },
          "documentId": {
            "type": "string"
          },
          "numberOfPages": {
            "type": "number"
          },
          "snippet": {
            "type": "string"
          }
        },
        "required": [
          "citationNumber",
          "matchScore",
          "fileName",
          "cloudinaryUrl",
          "documentId",
          "numberOfPages",
          "snippet"
        ]
      },
      "ChatResponseDto": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "response": {
            "type": "string"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatSourceDto"
            }
          }
        },
        "required": [
          "chatId",
          "success",
          "response",
          "sources"
        ]
      },
...
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@company.com"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "minLength": 8,
            "maxLength": 16
          }
        },
        "required": [
          "email",
          "name",
          "password"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "4dce2c35-c0e7-4f6a-8c4d-bd8108d2ed9c"
          },
          "email": {
            "type": "string",
            "example": "user@company.com"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          }
        },
        "required": [
          "id",
          "email",
          "name"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@company.com"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "minLength": 8,
            "maxLength": 16
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "RefreshResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "New access token (also set as HttpOnly cookie)"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "UpdateAuthDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}